finished bfs functions, also created subfolders for different functions
