+++ ant_farms/test_3
@@ -15,10 +15,8 @@
 9 13 8
 #nogeenrandomcomment
 10 13 9
-0-2
 0-1
 10-6
-2-3
 3-5
 3-4
 #coooooomment
@@ -29,5 +27,3 @@
 9-10
 7-10
 1-7
-1-2
-2-4
