+++ bfs_functions/sources
@@ -1,0 +1,2 @@
+SRCS += bfs_functions/bfs.c \
+bfs_functions/queue.c
