+++ hashtable_functions/list.c
