+++ hashtable_functions/set_links.c
