+++ hashtable_functions/sources
@@ -1,0 +1,3 @@
+SRCS += hashtable_functions/hash_table.c \
+hashtable_functions/list.c \
+hashtable_functions/set_links.c
