+++ input_functions/check_format.c
