+++ input_functions/read_input.c
@@ -6,7 +6,7 @@
 /*   By: igvan-de <igvan-de@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/24 14:28:43 by igvan-de       #+#    #+#                */
-/*   Updated: 2019/11/18 16:45:08 by igvan-de      ########   odam.nl         */
+/*   Updated: 2019/11/19 11:45:31 by igvan-de      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
