+++ input_functions/sources
@@ -1,0 +1,4 @@
+SRCS += input_functions/check_format.c \
+input_functions/lem_split.c \
+input_functions/read_input.c \
+input_functions/start_end.c
