+++ input_functions/start_end.c
