+++ sources
@@ -1,0 +1,2 @@
+SRCS = main.c \
+test_print.c
