+++ test_print.c
@@ -6,7 +6,7 @@
 /*   By: igvan-de <igvan-de@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/11/04 14:54:30 by igvan-de       #+#    #+#                */
-/*   Updated: 2019/11/18 13:27:09 by ygroenev      ########   odam.nl         */
+/*   Updated: 2019/11/19 11:12:34 by igvan-de      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,4 +62,29 @@ void    print_hash(t_table **table, size_t size)
         }
         i++;
     }
+}
+
+void	print_queue(t_queue *queue)
+{
+	t_queue *test;
+
+	test = queue;
+	while (test != NULL)
+	{
+		printf("test = %s\t", test->to->name);
+		printf("test->visisted = %d\t", test->to->visited);
+		printf("test->distence = %d\n\n", test->to->distance);
+		test = test->next;
+	}
+	printf("============================\n");
+	/*place pop_out_queu function here*/
+	printf("========AFTER POP=========\n");
+	test = queue;
+	while (test != NULL)
+	{
+		printf("test = %s\t", test->to->name);
+		printf("test->visisted = %d\t", test->to->visited);
+		printf("test->distence = %d\n\n", test->to->distance);
+		test = test->next;
+	}
 }
